implementando view de dados sensíveis

+++ app/src/main/java/com/douglasfj/myfinances/models/entities/CartoesEntity.java
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey;
 public class CartoesEntity {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "ID_CARTAO")
-    public int idCartao;
+    public Long idCartao;
 
     @ColumnInfo(name = "NOME_CARTAO")
     public String nomeCartao;
